pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Aasher Kamal", email = "aasherkamal786@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.4.1",
    "openai-agents>=0.0.4",
    "python-dotenv>=1.0.1",
]
